+++ src/dispatching/handler_factory.rs
@@ -1,6 +1,7 @@
 use dptree::{di::DependencyMap, Handler};
 
 /// Something that can construct a handler.
+#[deprecated(note = "Use the teloxide::handler! API")]
 pub trait HandlerFactory {
     type Out;